[MISC][WL] move custom component typings to separate file
