+++ src/components/custom/index.ts
@@ -1,1 +1,2 @@
 export * from "./filter";
+export * from "./types";
